+++ src/explorer/check_result.ts
@@ -1,3 +1,6 @@
+import { capitalize } from '../utils';
+
+
 export class CheckResult {
 	public code: string;
 	public provider: string;
@@ -5,14 +8,34 @@ export class CheckResult {
 	public filename: string;
 	public startLine: number;
 	public endLine: number;
+	public severity: string;
+	public summary: string;
+	public impact: string;
+	public resolution: string;
+	public docUrl: any;
 	constructor(
-		private result: any
+		result: any
 	) {
 		this.code = result.rule_id;
 		this.provider = result.rule_provider;
 		this.codeDescription = result.rule_description;
+		this.summary = result.description;
+		this.impact = result.impact;
+		this.resolution = result.resolution;
 		this.filename = result.location.filename;
 		this.startLine = result.location.start_line;
 		this.endLine = result.location.end_line;
+		this.severity = capitalize(result.severity);
+
+		if (result.links.length > 0) {
+			this.docUrl = result.links[0];
+		}
+	}
+}
+
+export class CheckSeverity {
+	public severity: string;
+	constructor(result: any) {
+		this.severity = capitalize(result.severity);
 	}
 }